requirements.txt
aiofiles
aiohttp
asyncio
cookidoo-api
fastapi[standard]
openai
pymilvus
python-dotenv